+++ src/extension.ts
@@ -36,7 +36,23 @@ export function activate(context: vscode.ExtensionContext) {
     }
   );
 
-  context.subscriptions.push(commandJest, commandMockito);
+  const commandEmpty = vscode.commands.registerCommand(
+    "angular-spec-generator.generateEmptyFile",
+    (event) => {
+      const fileHandler = new FileHandler();
+
+      fileHandler
+        .readFile(event.fsPath)
+        .then(fileHandler.getClassName)
+        .then((className) =>
+          fileHandler.createEmptyFile(event.fsPath, className)
+        )
+        .then(fileHandler.openFileInEditor)
+        .catch(vscode.window.showErrorMessage);
+    }
+  );
+
+  context.subscriptions.push(commandJest, commandMockito, commandEmpty);
 }
 
 // this method is called when your extension is deactivated