+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 All notable changes to the "angular-spec-generator" extension will be documented in this file.
 
+## v1.2.0
+
+- feat: Support generating a empty spec file
+- fix: call a service method in mockito service spec file
+
 ## v1.1.0
 
-- Support generating of mockito spec files
+- feat: Support generating of mockito spec files
   
 
 ## v1.0.0